requirements.txt
-i https://pypi.org/simple
aniso8601==10.0.0
annotated-types==0.7.0; python_version >= '3.8'
anyio==4.9.0; python_version >= '3.9'
blinker==1.9.0; python_version >= '3.9'
certifi==2025.1.31; python_version >= '3.6'
cffi==1.17.1; platform_python_implementation != 'PyPy'
charset-normalizer==3.4.1; python_version >= '3.7'
click==8.1.8; python_version >= '3.7'
cryptography==44.0.2; python_version >= '3.7' and python_full_version not in '3.9.0, 3.9.1'
distro==1.9.0; python_version >= '3.6'
filelock==3.18.0; python_version >= '3.9'
flask==3.1.0; python_version >= '3.9'
flask-cors==5.0.1; python_version >= '3.9' and python_version < '4.0'
flask-restful==0.3.10
flask-sqlalchemy==3.1.1; python_version >= '3.8'
fsspec==2025.3.0; python_version >= '3.8'
greenlet==3.1.1; python_version < '3.14' and platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
h11==0.14.0; python_version >= '3.7'
httpcore==1.0.7; python_version >= '3.8'
httpx==0.28.1; python_version >= '3.8'
huggingface-hub==0.29.3; python_full_version >= '3.8.0'
idna==3.10; python_version >= '3.6'
itsdangerous==2.2.0; python_version >= '3.8'
jinja2==3.1.6; python_version >= '3.7'
jiter==0.9.0; python_version >= '3.8'
markupsafe==3.0.2; python_version >= '3.9'
mpmath==1.3.0
networkx==3.4.2; python_version >= '3.10'
numpy==2.2.4; python_version >= '3.10'
nvidia-cublas-cu12==12.4.5.8; platform_system == 'Linux' and platform_machine == 'x86_64'
nvidia-cuda-cupti-cu12==12.4.127; platform_system == 'Linux' and platform_machine == 'x86_64'
nvidia-cuda-nvrtc-cu12==12.4.127; platform_system == 'Linux' and platform_machine == 'x86_64'
nvidia-cuda-runtime-cu12==12.4.127; platform_system == 'Linux' and platform_machine == 'x86_64'
nvidia-cudnn-cu12==9.1.0.70; platform_system == 'Linux' and platform_machine == 'x86_64'
nvidia-cufft-cu12==11.2.1.3; platform_system == 'Linux' and platform_machine == 'x86_64'
nvidia-curand-cu12==10.3.5.147; platform_system == 'Linux' and platform_machine == 'x86_64'
nvidia-cusolver-cu12==11.6.1.9; platform_system == 'Linux' and platform_machine == 'x86_64'
nvidia-cusparse-cu12==12.3.1.170; platform_system == 'Linux' and platform_machine == 'x86_64'
nvidia-cusparselt-cu12==0.6.2; platform_system == 'Linux' and platform_machine == 'x86_64'
nvidia-nccl-cu12==2.21.5; platform_system == 'Linux' and platform_machine == 'x86_64'
nvidia-nvjitlink-cu12==12.4.127; platform_system == 'Linux' and platform_machine == 'x86_64'
nvidia-nvtx-cu12==12.4.127; platform_system == 'Linux' and platform_machine == 'x86_64'
openai==1.68.2; python_version >= '3.8'
packaging==24.2; python_version >= '3.8'
pdfminer.six==20250324; python_version >= '3.8'
psycopg2-binary==2.9.10; python_version >= '3.8'
pycparser==2.22; python_version >= '3.8'
pydantic==2.10.6; python_version >= '3.8'
pydantic-core==2.27.2; python_version >= '3.8'
pytz==2025.2
pyyaml==6.0.2; python_version >= '3.8'
regex==2024.11.6; python_version >= '3.8'
requests==2.32.3; python_version >= '3.8'
safetensors==0.5.3; python_version >= '3.7'
setuptools==78.1.0; python_version >= '3.12'
six==1.17.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sniffio==1.3.1; python_version >= '3.7'
sqlalchemy==2.0.39; python_version >= '3.7'
sympy==1.13.1; python_version >= '3.9'
tokenizers==0.21.1; python_version >= '3.9'
torch==2.6.0; python_full_version >= '3.9.0'
tqdm==4.67.1; python_version >= '3.7'
transformers==4.50.1; python_full_version >= '3.9.0'
triton==3.2.0; platform_system == 'Linux' and platform_machine == 'x86_64'
typing-extensions==4.13.0; python_version >= '3.8'
urllib3==2.3.0; python_version >= '3.9'
werkzeug==3.1.3; python_version >= '3.9'
